+++ TCM_Hotel/nbproject/private/private.xml
@@ -3,7 +3,7 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/C:/Users/faari/git_pdc_files/tcm-hotel/TCM_Hotel/src/tcm/hotel/gui/WelcomeAppGui.java</file>
+            <file>file:/C:/Users/faari/git_pdc_files/tcm-hotel/TCM_Hotel/src/tcm/hotel/WelcomeAppGui.java</file>
         </group>
     </open-files>
 </project-private>
